Move TestHTTPClientBase to the helper module
