+++ marathon_acme/clients/tests/helpers.py
@@ -1,4 +1,48 @@
+from testtools import TestCase
+from testtools.twistedsupport import AsynchronousDeferredRunTest
+
+from treq.client import HTTPClient as treq_HTTPClient
+
+from twisted.internet.defer import DeferredQueue
+from twisted.web.server import NOT_DONE_YET
+
+from txfake import FakeHttpServer
+from txfake.fake_connection import wait0
+
 from uritools import urisplit
+
+
+class TestHTTPClientBase(TestCase):
+    # TODO: Run client tests synchronously with treq.testing tools (#38)
+    run_tests_with = AsynchronousDeferredRunTest.make_factory(timeout=0.1)
+
+    def setUp(self):
+        super(TestHTTPClientBase, self).setUp()
+
+        self.requests = DeferredQueue()
+        self.fake_server = FakeHttpServer(self.handle_request)
+
+        fake_client = treq_HTTPClient(self.fake_server.get_agent())
+        self.client = self.get_client(fake_client)
+
+        # Spin the reactor once at the end of each test to clean up any
+        # cancelled deferreds
+        self.addCleanup(wait0)
+
+    def handle_request(self, request):
+        self.requests.put(request)
+        return NOT_DONE_YET
+
+    def get_client(self, client):
+        """To be implemented by subclass"""
+        raise NotImplementedError()
+
+    def uri(self, path):
+        return '%s%s' % (self.client.url, path,)
+
+    def cleanup_d(self, d):
+        self.addCleanup(lambda: d)
+        return d
 
 
 class PerLocationAgent(object):
