+++ marathon_acme/clients/tests/test__base.py
@@ -1,26 +1,22 @@
-from testtools import ExpectedException, TestCase
+from testtools import ExpectedException
 from testtools.assertions import assert_that
-from testtools.matchers import (
-    Equals, Is, IsInstance,
-    MatchesStructure)
-from testtools.twistedsupport import (
-    AsynchronousDeferredRunTest, failed, flush_logged_errors)
+from testtools.matchers import Equals, Is, IsInstance, MatchesStructure
+from testtools.twistedsupport import failed, flush_logged_errors
 
 from treq.client import HTTPClient as treq_HTTPClient
 
 from twisted.internet import reactor
-from twisted.internet.defer import DeferredQueue, inlineCallbacks
+from twisted.internet.defer import inlineCallbacks
 from twisted.internet.task import Clock
 from twisted.web.client import Agent
 from twisted.web.http_headers import Headers
-from twisted.web.server import NOT_DONE_YET
 
-from txfake import FakeHttpServer
 from txfake.fake_connection import wait0
 
 from marathon_acme.clients._base import (
     HTTPClient, HTTPError, default_client, default_reactor, get_single_header,
     raise_for_status)
+from marathon_acme.clients.tests.helpers import TestHTTPClientBase
 from marathon_acme.clients.tests.matchers import HasRequestProperties
 from marathon_acme.tests.helpers import failing_client
 from marathon_acme.tests.matchers import HasHeader, WithErrorTypeAndMessage
@@ -105,39 +101,6 @@ class TestDefaultClient(object):
         agent.
         """
         assert_that(default_client(None, reactor), IsInstance(treq_HTTPClient))
-
-
-class TestHTTPClientBase(TestCase):
-    # TODO: Run client tests synchronously with treq.testing tools (#38)
-    run_tests_with = AsynchronousDeferredRunTest.make_factory(timeout=0.1)
-
-    def setUp(self):
-        super(TestHTTPClientBase, self).setUp()
-
-        self.requests = DeferredQueue()
-        self.fake_server = FakeHttpServer(self.handle_request)
-
-        fake_client = treq_HTTPClient(self.fake_server.get_agent())
-        self.client = self.get_client(fake_client)
-
-        # Spin the reactor once at the end of each test to clean up any
-        # cancelled deferreds
-        self.addCleanup(wait0)
-
-    def handle_request(self, request):
-        self.requests.put(request)
-        return NOT_DONE_YET
-
-    def get_client(self, client):
-        """To be implemented by subclass"""
-        raise NotImplementedError()
-
-    def uri(self, path):
-        return '%s%s' % (self.client.url, path,)
-
-    def cleanup_d(self, d):
-        self.addCleanup(lambda: d)
-        return d
 
 
 class TestHTTPClient(TestHTTPClientBase):
