+++ marathon_acme/clients/tests/test_marathon_lb.py
@@ -7,8 +7,8 @@ from txfake.fake_connection import wait0
 
 from marathon_acme.clients._base import HTTPError
 from marathon_acme.clients.marathon_lb import MarathonLbClient
+from marathon_acme.clients.tests.helpers import TestHTTPClientBase
 from marathon_acme.clients.tests.matchers import HasRequestProperties
-from marathon_acme.clients.tests.test__base import TestHTTPClientBase
 from marathon_acme.tests.matchers import HasHeader, WithErrorTypeAndMessage
